Move project over to .NET Core
There are several reasons for this.  First off, the project files for Vertigo were starting to get large and hard to read, so a refactor was in order anyway.

The second reason involves the fact that dependency management via a build-script can become rather cumbersome and hard to maintain.

The biggest issue, however, is the fact that we were building for both Mono and vanilla .NET and this was being determined by a build script and environment variable.  While there is nothing wrong with Mono, I feared that building with two separate compilers could lead to platform issues that could be hard to debug in the future.  By using .NET core, we still have compatibility between the "big three" (Windows, Mac, and Linux), and we have the benefit of having the compiler being built by one source instead of two.